refactor: navigation head와 side 분리
